refactor(repeater): 拆分 handlers 目录（P5）
__init__ 仅保留元数据、生命周期绑定与对外 re-export；
matcher/调度逻辑迁至 handlers/{message,ban,scheduler,lifecycle,helpers}。
行为不变，fanout 仍 lazy import 避免循环依赖。
