+++ src/plugins/repeater/handlers/__init__.py
@@ -1,0 +1,5 @@
+"""注册 repeater matcher 与调度任务。"""
+
+from . import ban, lifecycle, message, scheduler
+
+__all__ = ["ban", "lifecycle", "message", "scheduler"]
