+++ src/plugins/repeater/handlers/helpers.py
@@ -1,0 +1,62 @@
+"""复读插件消息后处理与禁言检测。"""
+
+from __future__ import annotations
+
+import time
+
+from nonebot import get_bot, logger
+from nonebot.adapters.onebot.v11 import Message, MessageSegment
+from nonebot.exception import ActionFailed
+
+from src.shared.utils.media_cache import get_image
+
+from ..model import Chat
+
+
+async def is_shutup(self_id: int, group_id: int) -> bool:
+    info = await get_bot(str(self_id)).call_api(
+        "get_group_member_info",
+        **{
+            "user_id": self_id,
+            "group_id": group_id,
+        },
+    )
+    flag: bool = info["shut_up_timestamp"] > time.time()
+
+    logger.info(f"bot [{self_id}] in group [{group_id}] is shutup: {flag}")
+
+    return flag
+
+
+async def post_proc(message: Message, self_id: int, group_id: int) -> Message:
+    new_msg = Message()
+    for seg in message:
+        if seg.type == "at":
+            try:
+                info = await get_bot(str(self_id)).call_api(
+                    "get_group_member_info",
+                    **{
+                        "user_id": seg.data["qq"],
+                        "group_id": group_id,
+                    },
+                )
+            except ActionFailed:
+                continue
+            nick_name = info["card"] or info["nickname"]
+            new_msg += f"@{nick_name}"
+        elif seg.type == "image":
+            cq_code = str(seg)
+            base64_data = await get_image(cq_code)
+            if base64_data:
+                new_msg += MessageSegment.image(file=base64_data)
+            else:
+                new_msg += seg
+        else:
+            new_msg += seg
+
+    if not await Chat.reply_post_proc(str(message), str(new_msg), self_id, group_id):
+        logger.warning(
+            f"bot [{self_id}] post_proc failed in group [{group_id}]: [{str(message)[:30]}] -> [{str(new_msg)[:30]}]"
+        )
+
+    return new_msg
